default options, fix orphan behvr, fix automatic removal, add confirmation for delete

+++ scripts/search_notes.sh
@@ -4,7 +4,7 @@ CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 source "$CURRENT_DIR/helpers.sh"
 
 if [[ -z "$1" ]]; then
-    action=$(printf "goto\nopen\ndelete\nswitch associativity\n[cancel]" | fzf --tmux --header='Select an action.')
+    action=$(printf "goto\nopen\ndelete\nswitch associativity\n[cancel]" | fzf --tmux $QN_FZF_OPTIONS --header='Select an action.')
 else
     action="$1"
 fi
@@ -26,13 +26,18 @@ elif [[ $action == "open" ]]; then
 elif [[ $action == "goto" ]]; then
     header="Open a note (in its associated location)"
 elif [[ $action == "switch associativity" ]]; then
-    associativity=$(printf "window\nsession\nglobal\n[cancel]" | fzf --tmux --header='Select an associativity.')
+    if [ -f "$STATE_FILE" ]; then
+        read -r cur_assoc < "$STATE_FILE"
+    else
+        cur_assoc = $(default_associativity)
+    fi
+    associativity=$(printf "window\nsession\nglobal\n[cancel]" | fzf --tmux $QN_FZF_OPTIONS --header="Select an associativity (current: $cur_assoc)")
     [[ "$associativity" == "[cancel]" || -z "$associativity" ]] && exit
     printf '%s\n' "$associativity" > "$STATE_FILE"
     exit
 fi
 
-mapfile -t target_map < <(printf "$out" | fzf --tmux $args --header="$header" --preview="$CURRENT_DIR/.preview_note {} $nd")
+mapfile -t target_map < <(printf "$out" | fzf --tmux $QN_FZF_OPTIONS $args --header="$header" --preview="$CURRENT_DIR/.preview_note {} $nd")
 [[ ${#target_map[@]} -eq 0 ]] && exit
 
 # Extract filenames
@@ -43,6 +48,8 @@ for t in "${target_map[@]}"; do
 done
 
 if [[ $action == "delete" ]]; then
+    confirm=$(printf "no\nyes\n" | fzf --tmux $QN_FZF_OPTIONS --header="Are you sure you want to delete ${targets[*]}?")
+    [[ "$confirm" == "no" || -z "$confirm" ]] && exit
     printf "%s\n" "${targets[@]}" | xargs -I{} rm "$nd/{}"
     exit
 fi
